requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
flask
flask-cors
pillow
requests